refactor: revert back to model interface controller format
