+++ src/api/attendance/attendance.interface.ts
@@ -1,0 +1,10 @@
+import {ObjectId} from 'bson';
+
+interface Attendance {
+  _id: ObjectId;
+  userId: ObjectId;
+  checkedIn: Date;
+  checkedOut: Date;
+};
+
+export default Attendance;
